+++ src/main/java/campaign/service/dto/RewardDTO.java
@@ -18,9 +18,9 @@ public class RewardDTO {
 
     private Integer prizeType;
 
-    private Integer prizeValue;
+    private Float prizeValue;
 
-    private Integer numOfPrize;
+    private Float numOfPrize;
 
     private Integer released;
 
@@ -120,19 +120,19 @@ public class RewardDTO {
         this.prizeType = prizeType;
     }
 
-    public Integer getPrizeValue() {
+    public Float getPrizeValue() {
         return prizeValue;
     }
 
-    public void setPrizeValue(Integer prizeValue) {
+    public void setPrizeValue(Float prizeValue) {
         this.prizeValue = prizeValue;
     }
 
-    public Integer getNumOfPrize() {
+    public Float getNumOfPrize() {
         return numOfPrize;
     }
 
-    public void setNumOfPrize(Integer numOfPrize) {
+    public void setNumOfPrize(Float numOfPrize) {
         this.numOfPrize = numOfPrize;
     }
