+++ src/main/java/campaign/web/rest/vm/RewardVM.java
@@ -14,9 +14,9 @@ public class RewardVM {
 
     private Integer prizeType;
 
-    private Integer prizeValue;
+    private Float prizeValue;
 
-    private Integer numOfPrize;
+    private Float numOfPrize;
 
     private Integer released;
 
@@ -68,19 +68,19 @@ public class RewardVM {
         this.prizeType = prizeType;
     }
 
-    public Integer getPrizeValue() {
+    public Float getPrizeValue() {
         return prizeValue;
     }
 
-    public void setPrizeValue(Integer prizeValue) {
+    public void setPrizeValue(Float prizeValue) {
         this.prizeValue = prizeValue;
     }
 
-    public Integer getNumOfPrize() {
+    public Float getNumOfPrize() {
         return numOfPrize;
     }
 
-    public void setNumOfPrize(Integer numOfPrize) {
+    public void setNumOfPrize(Float numOfPrize) {
         this.numOfPrize = numOfPrize;
     }
